Japan2017
Instead of using an excel database, this file has the database stored
in an array of objects that have an ArrayList of PVectors as its
attribute. The input is then compared to each of these ArrayLists to
find the best vector fit. It theoretically should work but it doesn’t.